aqua/requirements.txt
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
Django==1.9
django-bootstrap3==11.1.0
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-js-asset==1.2.2
geojson==2.4.1
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.0
pyowm==2.10.0
pyparsing==2.4.2
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
